requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
asyncpg==0.31.0
pydantic==2.12.0
pydantic-settings==2.13.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.20
sib-api-v3-sdk==7.6.0

